add more users

+++ src/actions/post.action.ts
@@ -1,14 +1,17 @@
 'use server'
 
+import { authOptions } from '@/app/api/auth/[...nextauth]/route'
 import { TPost } from '@/constants/types'
 import { connectToDatabase } from '@/lib/utils'
 import Post from '@/models/Post'
 import { getServerSession } from 'next-auth'
 import { revalidatePath } from 'next/cache'
 
-export async function createPost(post: Omit<TPost, 'postId'>) {
+export async function createPost(
+  post: Omit<TPost, 'postId' | 'createdBy' | 'createdAt'>
+) {
   try {
-    const session = await getServerSession()
+    const session = await getServerSession(authOptions)
     if (!session?.user?.name) {
       throw new Error('Not logged in')
     }
@@ -18,6 +21,7 @@ export async function createPost(post: Omit<TPost, 'postId'>) {
     await Post.create({
       postId: crypto.randomUUID(),
       ...post,
+      createdBy: session?.user?.name,
     })
 
     console.log('Post created:', post)
@@ -27,32 +31,60 @@ export async function createPost(post: Omit<TPost, 'postId'>) {
   }
 }
 
-export async function updatePost(postId: string, post: Omit<TPost, 'postId'>) {
+export async function updatePost(
+  postId: string,
+  newPost: Omit<TPost, 'postId' | 'createdAt' | 'createdBy'>
+) {
   try {
-    const session = await getServerSession()
+    const session = await getServerSession(authOptions)
     if (!session?.user?.name) {
       throw new Error('Not logged in')
     }
 
     await connectToDatabase()
 
-    await Post.updateOne({ postId }, { $set: post })
+    const post = await Post.findOne({ postId })
+
+    if (!post) {
+      throw new Error('Post not found')
+    }
+
+    if (session.user.name !== post.createdBy && session.user.role !== 'admin') {
+      throw new Error('Nemáš práva na úpravu tohoto příspěvku')
+    }
+
+    await Post.updateOne({ postId }, { $set: newPost })
 
     console.log('Post updated:', postId)
     revalidatePath('/')
   } catch (error) {
     console.error('Error updating post:', error)
+    if (error instanceof Error) {
+      throw new Error(error.message)
+    } else {
+      throw new Error('Chyba při mazání uživatele')
+    }
   }
 }
 
 export async function deletePost(postId: string) {
   try {
-    const session = await getServerSession()
+    const session = await getServerSession(authOptions)
     if (!session?.user?.name) {
       throw new Error('Not logged in')
     }
 
     await connectToDatabase()
+
+    const post = await Post.findOne({ postId })
+
+    if (!post) {
+      throw new Error('Post not found')
+    }
+
+    if (session.user.name !== post.createdBy && session.user.role !== 'admin') {
+      throw new Error('Nemáš práva na smazání tohoto příspěvku')
+    }
 
     await Post.deleteOne({ postId })
 
@@ -60,6 +92,11 @@ export async function deletePost(postId: string) {
     revalidatePath('/')
   } catch (error) {
     console.error('Error deleting post:', error)
+    if (error instanceof Error) {
+      throw new Error(error.message)
+    } else {
+      throw new Error('Chyba při mazání příspěvku')
+    }
   }
 }
